--- py2neo/graphio/objects/relationship.py
@@ -1,44 +1,0 @@
-#!/usr/bin/env python
-# -*- encoding: utf-8 -*-
-
-# Copyright 2011-2020, Nigel Small
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-import logging
-
-
-log = logging.getLogger(__name__)
-
-
-class Relationship(object):
-
-    TYPE = None
-
-    def __init__(self, start_node_labels, end_node_labels, start_node_properties,
-                 end_node_properties, properties):
-
-        self.start_node_labels = start_node_labels
-        self.end_node_labels = end_node_labels
-        self.start_node_properties = start_node_properties
-        self.end_node_properties = end_node_properties
-        self.properties = properties
-        self.object_type = self.TYPE
-
-    def to_dict(self):
-        return {
-            'start_node_properties': self.start_node_properties,
-            'end_node_properties': self.end_node_properties,
-            'properties': self.properties
-        }
