+++ py2neo/graphio/queries.py
@@ -98,8 +98,8 @@ def nodes_merge_unwind(labels, merge_properties, property_parameter=None):
     return q
 
 
-def query_create_rels_unwind(start_node_labels, end_node_labels, start_node_properties,
-                             end_node_properties, rel_type, property_identifier=None):
+def _query_create_rels_unwind(start_node_labels, end_node_labels, start_node_properties,
+                              end_node_properties, rel_type, property_identifier=None):
     """
     Create relationship query with explicit arguments.
 
@@ -144,7 +144,7 @@ def query_create_rels_unwind(start_node_labels, end_node_labels, start_node_prop
     return q
 
 
-def query_merge_rels_unwind(start_node_labels, end_node_labels, start_node_properties,
+def _query_merge_rels_unwind(start_node_labels, end_node_labels, start_node_properties,
                              end_node_properties, rel_type, property_identifier=None):
     """
     Merge relationship query with explicit arguments.
@@ -190,3 +190,38 @@ def query_merge_rels_unwind(start_node_labels, end_node_labels, start_node_prope
     q += "SET r = rel.properties RETURN count(r)\n"
 
     return q
+
+
+def _params_create_rels_unwind_from_objects(relationships, property_identifier=None):
+    """
+    Format Relationship properties into a one level dictionary matching the query generated in
+    `query_create_rels_from_list`. This is necessary because you cannot access nested dictionairies
+    in the UNWIND query.
+
+    UNWIND { rels } AS rel
+    MATCH (a:Gene), (b:GeneSymbol)
+    WHERE a.sid = rel.start_sid AND b.sid = rel.end_sid AND b.taxid = rel.end_taxid
+    CREATE (a)-[r:MAPS]->(b)
+    SET r = rel.properties
+
+    Call with params:
+        {'start_sid': 1, 'end_sid': 2, 'end_taxid': '9606', 'properties': {'foo': 'bar} }
+
+    :param relationships: List of Relationships.
+    :return: List of parameter dictionaries.
+    """
+    if not property_identifier:
+        property_identifier = 'rels'
+
+    output = []
+
+    for r in relationships:
+        d = {}
+        for k, v in r.start_node_properties.items():
+            d['start_{}'.format(k)] = v
+        for k, v in r.end_node_properties.items():
+            d['end_{}'.format(k)] = v
+        d['properties'] = r.properties
+        output.append(d)
+
+    return {property_identifier: output}
